Add files via upload
This file tests the load cell.  It tares the load cell and then runs a loop where it only prints the value of the load cell reading until it detects more than 200 grams, it then appends the latest reading to a list.  if it detects less than 200 grams it goes back to just printing the load cell weight.  When somebody hits the keyboard interrupt, ctrl+c, it cleans up the GPIO pin assignment and calculates the average of all the weight readings over 200 grams.